Renamed utility to time_utility
